// src/types.ts
export type OneBotMessageSegment =
  | { type: "text"; data: { text: string } }
  | { type: "image"; data: { file: string; url?: string; subtype?: number; summary?: string } }
  | { type: "at"; data: { qq: string } }
  | { type: "reply"; data: { id: string } }
  | { type: "face"; data: { id: number } }
  | { type: "record"; data: { file: string } }
  | { type: "video"; data: { file: string } }
  | { type: "music"; data: { type: string; id: string } }
  | { type: "tts"; data: { text: string } };

export type OneBotMessage = OneBotMessageSegment[] | string;

export type OneBotEvent = {
  time: number;
  self_id: number;
  post_type: string;
  meta_event_type?: string;
  message_type?: "private" | "group";
  sub_type?: string;
  message_id?: number;
  user_id?: number;
  group_id?: number;
  message?: OneBotMessage | string;
  raw_message?: string;
  sender?: {
    user_id: number;
    nickname: string;
    card?: string;
    role?: string;
  };
  target_id?: number;
  notice_type?: string;
};

export interface ImageMessageOptions {
  file: string;
  subtype?: number;
  summary?: string;
}

export interface EmojiMessageOptions {
  file: string;
  summary?: string;
}